+++ app/core/models/__init__.py
@@ -4,8 +4,3 @@ from . import ticket_comment_base    # pylint: disable=W0611:unused-import
 from . import ticket_comment_solution    # pylint: disable=W0611:unused-import
 from . import ticket_comment_task    # pylint: disable=W0611:unused-import
 from .ticket_dependencies import TicketDependency    # pylint: disable=W0611:unused-import
-
-from .model_notes import ModelNotes    # pylint: disable=W0611:unused-import
-from .notes import Notes    # pylint: disable=W0611:unused-import
-from .ticket.ticket_category_notes import TicketCategoryNotes    # pylint: disable=W0611:unused-import
-from .ticket.ticket_comment_category_notes import TicketCommentCategoryNotes    # pylint: disable=W0611:unused-import