+++ app/devops/models/__init__.py
@@ -1,10 +1,5 @@
 from . import check_ins    # pylint: disable=W0611:unused-import
 from . import git_group    # pylint: disable=W0611:unused-import
-from . import git_group_notes    # pylint: disable=W0611:unused-import
 from . import git_repository    # pylint: disable=W0611:unused-import
 from . import feature_flag    # pylint: disable=W0611:unused-import
-from . import feature_flag_notes    # pylint: disable=W0611:unused-import
 from . import software_enable_feature_flag    # pylint: disable=W0611:unused-import
-
-from .git_repository.github_notes import GitHubRepositoryNotes    # pylint: disable=W0611:unused-import
-from .git_repository.gitlab_notes import GitLabRepositoryNotes    # pylint: disable=W0611:unused-import