+++ app/project_management/models/__init__.py
@@ -1,4 +1,0 @@
-from .project_notes import ProjectNotes    # pylint: disable=W0611:unused-import
-from .project_type_notes import ProjectTypeNotes    # pylint: disable=W0611:unused-import
-from .project_state_notes import ProjectStateNotes    # pylint: disable=W0611:unused-import
-from .project_milestone_notes import ProjectMilestoneNotes    # pylint: disable=W0611:unused-import